qa: qa/1671 to test replaying a corrupted multi-archive set
RHBZ#1673053

qa/archives/multi-corrupted/20150508.11.50.0.xz is deliberately
corrupted due to a truncated log record (missing 4 byte trailer
in the last log record).

This archive can be replayed in non-multi mode, but loops forever
in multi-archive replay without making forward progress. e.g.
pmrep -a qa/archives/multi-corrupted kernel.percpu.cpu.user
That is the subject of RHBZ#1673053.

A work-in-progress patch to libpcp/src/interp.c is RFC and not
included in this commit. The patch checks for PM_ERR_LOGREC
error returns from some internal functions in libpcp/src/interp.c
that were otherwise ignored and assumed to be PM_ERR_EOL when
replaying across the corruption.

qa/1671 is expected to fail until a fix is committed.