qa/412: refactor arithmetic
On some platforms (*BSD in particular), awk's arithmetic is pretty lame
and does not deal at all well with values larger than a signed 32-bit
integer.

Pull the arithmetic out of awk and give it to bc(1) who can be
trusted to do the job.

Also some problems with the archive start time extraction which was
making the test unstable on slower VMs.