pmdabpftrace: don't use cached instances
if a dict is used for indom.set_instances, the instance names get cached
using the pmdaCacheOp functionality. However, for bpftrace maps, the keys
should not be cached. For example, at some point the metric
bpftrace.scripts.script100.data.bytes can contain a histogram, and at another
time the same metric can contain an indom with process names as the instance
names.