qa/1141: notrun if pcpupstream module not loaded
If we're on a system with selinux tools installed, but the pcpupstream
module is not loaded, then we run the risk of this test failing because
some other selinux dependency is not satisfied.  For example on SUSE
SLES12 (vm30), I'm seeing this:

Failed to resolve typeattributeset statement at /var/lib/selinux/minimum/tmp/modules/400/pcpupstream/cil:1
semodule:  Failed!

So if we can't load the official pcpupstream module, then there is no
hope for qa/1141, so notrun it.