libpcp,libpcp_pmda: fortify pmdaCache against NULL instance names
A while back Marko reported pminfo segfaults involving bcc metrics
when used to instrument Java12. This turned out to be NULL instance
names being stored in the pmdaCache file (which is actually a bug
in pmdabcc).

In defense of that scenario, fortify libpcp_pmda and libpcp
against NULL instance names being read in from the indom cache;
rather than segfault, return an appropriate error.

Also update qa/1152 to test and verify the fix. The pmdabcc bug
will be fixed separately now that it's easier to debug.