qa/870: relax checks for pmlogger control files
In the normal thrust and parry of QA, it is inevitable that some
pmlogger instances will be smacked out of existance in ways that may
prevent orderly cleanup.

If we find turdlets in $PCP_TMP_DIR/pmlogger or $PCP_RUN_DIR that
belong to a pmlogger process that is no longer running, we now treat
these as Warnings that are sent to $seq.full, rather than Errors that
are emitted on stdout.