misc: pander to gcc warning insanity
We want to use strtol() and strtod() to parse a string for errors.
With the flags we use for gcc this is not possible without gcc
emitting warnings.

To see how to fix this, check the commit, but be warned you'll
want to vomit.  If anyone has a less ugly hack, to achieve the
same result, please let me know.