python: refresh indom buffers if need_refresh flag is set in refresh cb
Original problem:
When updating uncached instance domains in the refresh callback of the
PMDA, the list of instances was corrupt (one or more instances were
missing).

Cause:
The it_set pointer in libpcp_pmda pointed to a ctypes array which got
garbage collected by the python runtime.

Resolution:
Set the need_refresh flag in set_list_instances, check if this flag is
set after calling the refresh callback of the PMDA, and copy the indom
buffers (if required)

Alternative solution:
By using the refresh_metrics callback and setting needs_refresh in the
PMDA, this problem could also be solved. However, the PMDA shouldn't
need to take care of implementation details like the need_refresh flag,
and it should be possible to update instance domains from the
refresh(cluster) callback as well.