qa: remake 1362 for postgresql and add setup/init comments
qa/1362 is _notrun if the postgresql server is not installed,
initialized and the service has successfully started. For details
and instructions, see the (new) comments at the top of qa/1362. In
particular, after a major system upgrade (e.g. f29 to f30), a once-off
DB migration may be necessary - see comments in the test.

Also remade 1362.out for the minor log message change after Nathan's
last round of changes (Warnings became Info messages).