libpcp_pmda: set context in store callback
All callbacks in callback.c called __pmdaSetContext to store the current
context number, except the store callback, which was a no-op previously.

This commit sets the context number in the store callback, and calls the
store callback from the Python wrapper. In most cases the context was
set without this change, as a call to pmdaDesc set the context properly.
However, there is no guarantee that pmdaDesc is called before store.