QA 1131: adjust for different Python versions
Before this patch the test was failing on CentOS/RHEL 6 Python 2.6.

With -E we want to see the output values as numbers, not strings,
the exact precision is not that important so decrease precision to
avoid issues with different Python version. In theory this change
could be reverted one day when CentOS/RHEL 6 becomes unsupported.