libpcp_web: harden null string value handling in webgroup fetches
The bpftrace overview dashboard in Grafana was hitting a situation where
it would have null-valued bpftrace metrics initially for a short time -
these were not being handled correctly in the library.  Tested using the
Grafana interface as there's no other way to reproduce this I'm aware of
at this stage.