pmlogrewrite: plug major memory leak
We were keeping all the PDU buffers because they were never being
unpinned.  This may explain some folklore around long running pmlogger
daily tasks with big archives.

Simple fix.

And added conditional call to __pmFindPDUBuf(-1) before exit to dump
the state of the PDU buffer pool to debug this.

And new qa/1149 to demonstrate the problem and prove the fix works.

Passes all of -g pmlogrewrite QA, so I'm sure this is safe.