pmproxy: send only one fetch response
pmWebGroupFetch() calls webgroup_fetch_names() which calls
webgroup_fetch() which calls on_pmwebapi_done()
However, at the end of pmWebGroupFetch(), on_pmwebapi_done() is called
also.

This results in (trying to) send the actual fetch response *and* a
generic {context:XXX, success:true} response, and for HTTP/1.0 clients to
close the connection twice, which results in a libuv assert.

This commit removes the call to on_pmwebapi_done in the webgroup_fetch()
function, and instead returns the error message and status code via
a parameter and return value (like the other webgroup_* functions).