pmdalinux: fix segfault in interrupts code with large CPU counts
Fix problems in buffer sizing and PMID item-number wrapping when
presented with more than 1024 /proc/interrupts interrupt lines.

Update qa/886 to exercise these fixes.  Note the output of this
test is now compressed due to the size of the pminfo output when
run on machines of this size - its decompressed on the QA host,
just prior to running.

This resolves Red Hat BZ #1755069