qa/admin/other-packages/require: need python2-devel for Fedora 28
This seems to depend on both python2 and python3 being installed.
I know Fedora 28 is at EOL, but I suspect this issue is present in
other platforms where multiple versions of Python are installed,
so making the change here provides a pattern on how to fix it if
the issue reappears elsewhere.