pmdaroot: improve handling of container indom updates
Fixes Jason, Mark and I have been fine-tuning during the week,
improving the handling of container instances that are not active
initially when scanning, but become so later.  A container flag
is added to indicate when an initial heuristic was used for a
cgroup name rather than an exact-match from filesystem traversal
in the cgroup filesystem.

Extend the qa/761 coverage with sample docker state from Netflix.