src/libpcp/src/mk.pmdbg: rework last change
The change made in commit 9f267705 for Fedora 31 to workaround a change
in awk behaviour (why, oh, why did anyone think changing awk in this way
was a good idea), unfortunately breaks with older versions of awk, e.g.
on RHEL 6.10.

Rework the change to avoid the \{ in the pattern (which is 100% legal
and should not generate a warning) and replace this with [{].

Sigh.