pmdaproc: support for kernel cgroups v2 (and systemd using same)
Fedora31 has fully enabled the kernels newer style of cgroup (v2),
along with the systemd changes to start using these cgroups.  This
changes the way we need to detect cgroups completely - the unified
form is also not used in this setup (which provided v1 and v2 side
by side, but crippling v2 in the process).

So, update the logic in pmdaproc (cgroup.* metrics and per-process
metrics within a container) so it works with the new world order.
Add in new cgroup.cpu.*, cgroup.pressure.*, cgroup.memory and per-
device cgroup.io.* metrics while we're at it.

Resolves https://github.com/performancecopilot/pcp/issues/381
Resolves https://github.com/performancecopilot/pcp/issues/654
Resolves https://github.com/performancecopilot/pcp/issues/659