pcp-dstat: fix handling of very large numbers of disk devices
Update the generic pmconfig code to be more dynamic in the way it
handles the fetchgroup maximally-sized-arrays requirement.  We're
now tracking the largest instance list size seen, then using that
instead of hard-coding these limits.

This has two benefits - it reduces the amount of memory allocated
in the default cases, and ramps up at runtime to whatever size we
need to use.  Fixes pcp-dstat calling convention which previously
was using the default 1024 limit and has been observed to fail on
systems with large numbers of disks.  Test qa/1207 exercises this
case specifically, but the solution is applicable for interfaces,
CPUs, etc also.

Resolves Red Hat BZ #1779419.