qa, libpcp_web: fix leak scraping sample.secret.foo.bar.max.redirect
Due to the quirks of this metric, which 'redirects' to a PMID in
the pmcd PMDA along with its sampledso counterpart, name resolution
(esp. via pmNameAll) is problematic.

Change the code to also associate the original name with the new
in-memory metric structure so we can correctly report and free up
memory later.

Another step toward resolving qa/1695 valgrind issues (one left).