pmwebapi: do not free the context level labelset in pmSeriesDiscoverLabels
Do not free context labelset after dup'ing it in pmSeriesDiscoverLabels.
This labelset is created and tracked by the libpcp_web discover API and
causes a double-free when the tracked archives (and their labelsets)
are later purged when the archives are culled or compressed/deleted.

This is a minimal patch to resolve PCP GH#800 at this stage.
Resolves https://github.com/performancecopilot/pcp/issues/800