pcp2xml: adjust escaping rules
Rename the helpers to better reflect their function.

Do not escape single-quotes (') which is not mandatory, this will
make output much more readable when labels are supported. Handle
the special case of ']]>' while at it.

Also fix the "hardening" of the helpers to properly deal with
empty strings, never return None but "" instead, adjust QA accordingly.