pmrep/pcp2xxx: handle corner case of initially missing instances
If no instances are present during pmrep et al startup, the metric
was mistakenly interpreted as a single-valued (flat) metric. Handle
this corner case which had been present since the introduction of
pmrep(1), even if it seems to be extremely rare a library/module
should nevertheless always do the right thing. The cost is two
int vs int checks for each metric and instance on each fetch, does
not seem like overkill.

QA 1168 added to test this, fails without the pmconfig.py patch,
passes with it.