pmrep: handle corner case of dis/appearing instances
In a rare (and in many cases, practically impossible) case of one
instance disappearing from one metric but an instance with the
same name appearing for another metric pmrep(1) would not detect
this change in instances properly. Fix this by keeping record of
previous instances more accurately.

Add a generated archive to test this. The archive also shows the
nature of the issue: e.g., instance x disappears from disk.dev.read
but right at the next fetch appears for (and only for) another
metric like disk.dev.write.