libpcp_web: fix memory leak when discovery enabled without redis-server
When operating in async mode, the Redis client code buffers up commands
until the connection is available.  This can lead to memory use blowout
when we're unable to transfer that data in reasonable time.

There may be no Redis server available at all, intentionally - in this
case (as well as other persistent connection loss) we need to discard
buffered commands.  Add a flag indicating connection state and prevent
this buffering of commands from building up in these cases.

New test qa/1573 is added to verify this.

Resolves https://github.com/performancecopilot/pcp/issues/804