libpcp_web: improve handling of metric 'cached' field
With Marks recent discovery changes, we now have a single context
structure with a single archive - this means an earlier behaviour
of dup'ing some in-core structures no longer happens and metadata
writes (to Redis) could get skipped.  To prevent this start using
the 'cached' field in the metric structure as intended.