python: fix pmParseUnitsStr for Python 2, fixes QA 702 and 707
The function raised an exception when called with an unicode object.
This commit introduces a new text_type compatibility variable which
points to str in Python3, and unicode in Python2. Python 2.6 and 2.7
include bytes as a builtin alias of str, which makes the code now
functional with both Python 2 and 3.