libpcp_web: handle dynamic metadata changes in log discovery
Make use of the new libpcp interfaces for adding new metadata
on-the-fly as we detect it in libpcp_web discovery.  While in
there, make pmseries queries that provide only 'count' (among
the time window parameters) return N-most recent values as it
is a/ much more likely to be useful, and b/ the users intent.

Update several QA tests to exercise these changes.