pmdalinux: convert uptime and idletime metrics to double precision
Previously we exported kernel.all.uptime and idletime as unsigned
32bit quantities, discarding any sub-second time component from the
kernel.  However calculations involving subtracting uptime from the
start time of a process (in milliseconds) have been observed to go
negative as a result of the lost fraction component.

The kernel exports a timespec (with 64bit seconds) converted into
floating point before exporting, so preserve this information now
in the PCP metrics.

QA test 1416 is added to exercise the archive rewriting rules.