qa: fix 1661 to work with multiple pmlogger/farm and add _check_series check
Add _check_series check to the test which will _notrun on
platforms that do not support pmseries, e.g. no libuv.
This might fix Ken's issues with this test on opensuse,

Also fix the curvol check to work on systems with multiple
pmloggers running - just check the current logvol for one
of the pmloggers. This should fix Nathan's issues with this
test.