build: add static checking for man pages, for CI
Use mandoc and lexgrog to perform some man page validity checking.
This is run in the (optional) 'make check' phase of the build.

Some configure/builddefs tidyup of the cppcheck invocation while I
was in this neck of the woods.

Lots of errors and warning level issues being reported, so not all
man pages are checked yet - this just gets us started (any/all help
welcome!).