src/libpcp/src/derive_fetch.c: fix obscure mem leak
To trigger this needed pmFetch() with both derived metrics and regular
metrics and the regular metrics included at least one metric with
a valfmt == PM_VAL_SPTR and we go down the don't rewrite path in
__dmpostfetch() ... this usually means a metric with a value that is
a set of event records.

Only known trigger was qa/1695, but only on some hosts.

Sigh.