qa: remove assumption from test 1661 about local archives
Test 1661 was failing here due to the presence of non-primary archive
logs (not loggers, which it was catering for, but pre-existing logs).
While fixing that, I needed extra diagnostics in the .full file, and
noticed we can use pmlc to do the log rotation reliably, rather than
using signal/sleep.