python/pmapi: plug memleaks in labelsets API calls, plus QA
As Marko found, the python pmapi binding has been leaking
memory in the labels lookup API related methods. This was
found to be due to arrays of pmLabelSet structures allocated
by the libpcp calls not being free'd after those arrays are
converted into dicts in the python binding (they are not
ref counted and so not garbage collected).

qa/1490 is a new test that exercises the fix, using a simple
test program in qa/src/labelsets_memleak.python.