find-filter: (new) find-grained filter for files based on mtime or ctime
find-filter is a new helper filter (installed in $PCP_BINADM_DIR) that
is intended to be used in conjunction with find(1) to produce better
than one day selection of files, based on their last modification time
(or creation time).

The implementation is platform-independent and has resolution to a
minute relative to the current time.  This avoids the "day rounding"
algorithms of find(1)'s -mtime that differ between platforms, and
avoids the non-standard -mmin option.

This will be used on pmlogger_daily to provide shorter and semantically
sounder windows for culling and compressing operations.