libpcp: add portable service manager notify API functions
Add two additional auxserver API functions in libpcp for daemons
to notify their service manager (usually systemd, but can be any
or even none) that they have started and now ready, along with
the PID of the main process. These functions are intended to be
portable in that the platform dependent code is buried in the
library and so no conditional code is needed in the daemons
themselves. For platforms with no service manager (other than
init), these calls reduce to a cheap noop.

The advantage of using a notify API is that the daemons can choose
to fork or daemonize or exec etc (or not do so) without the service
manager having to try and track or guess the main PID of the service.
Another advantage is when a start/rc script (if any) launches the
daemon, it need not necessarily busy wait for the service to begin.
The daemon knows when it's ready and notifies the service manager
(using this API).

Diagnostics will be logged if the pmDebugOptions.services flag
is set. Also see qa/1566 and src/test_service_notify.c (in the
next commit), along with new test qa/1566 and the documentation in
the new man page __pmServerNotifyServiceManagerReady(3i).