src/libpcp/src/auxserver.c: change guards around libsystemd uses
Use HAVE_SYSTEMD_SD_DAEMON_H for header and call to sd_notify().

Add code to issue warning if __pmServerNotifySystemd() is called on
a system where HAVE_SYSTEMD is true but HAVE_SYSTEMD_SD_DAEMON_H is
false ... this is possibl, for example, on Debian-based systems where
we are using systemd, but the libsystemd-dev package was not installed
when the pcp package was built, so we have no headers and no library
to provide sd_notify().

Also make the warning unconditional (not behind a pmDebugOptions.services
guard), when __pmServerNotifySystemd() is called for a non-systemd
environment (which should not happen).