pmlogger: write out Latest archive folio when running as a service daemon
On startup, just prior to entering the main loop, write out the
"Latest" archive folio with the current archive basename. This is
ONLY done if this pmlogger is running as a service daemon (including
if the process has been reexec'd) and the pmlogger cwd is prefixed
with $PCP_ARCHIVE_DIR. If either of those conditions is not true,
then the folio is not written.

Note: this task used to be done by the logutil scripts. Until that
functionality is removed, the logutil service scripts will write
out their own version of the Latest folio (approx 2 seconds after
pmlogger).