src/pmlogger/src/pmlogger.c: change duplicate archive name avoidance strategy
When pmlogger is launched repeatedly from systemd (or anywhere using
the strftime pattern for the archive name), there is a chance that more
than one invocation happens inside a minute, in which case the %0M at
the end of the strftime pattern is not enough to guarantee uniqueness.

Previously we we just pretending that we were starting +1 or +2 or
+3 miniutes later, but this approach is doomed in the presence of
rapid restarting, like in QA.

This commit reinstates the scheme from pmnewlog(1) (now deprecated)
and uses suffixes like "", "-00", "-01", ... to force the output
archive base name to be unique.

In the process this means the real archive name maybe longer
than the strftime pattern, so drop the argv[] munging, which on
reflection appears to be (a) non-portable, (b) not reliable, and (c)
not necessary.