src/pmlogger/pmlogger_check.sh: fortify test for archive creation
Previously we knew the base name of the newly created archive, but
recent changes have meant that this decision is made by pmlogger and
so this script is left to deduce the base name.

If pmlogger does not start (like on purpose in some QA tests), then
the archive is not created and the logic here was a bit busted.