pmlogger: refactor -NN archive name collision avoidance
The pre-test with access() is insufficient.

Need to move __pmLogCreate() inside the -NN loop and break when this
succeeds.

A bunch of knock-on effects of this, as the real archive name is not
known until the __pmLogCreate() succeeds, and may be different to the
archive name after the initial strftime() expansion.