qa/check.callback.sample: tweak PMNS check
For reasons that are not yet understood, the romance between the pmproxy
daemon and the mmv PMDA wearing its pmproxy PMDA disguise can sour.
When this happens, the pmproxy.pid metric is left in the PMNS but
the mmv (aka pmproxy) PMDA returns PM_ERR_PMID, so teak the filter
to not raise this as a callback detected error.