src/pmproxy/src/pmproxy.c: make logfile name unique with -Dappl1
One(more help for debugging, with -Dappl1, the process' pid is embedded
in the string ".<pid>" and this is stitched into) the logfile name
to help make it unique.  This helps when multiple pmproxy instances
are involved in a QA run.

Beware, in a default setup this will pollute $PCP_LOG_DIR/pmproxy
with lots of pmproxy.NNNN.log files and these will need to be cleaned
up manually.