qa/1190: fix botch in last commit
Changes to the quoting logic were botched in the last commit
... making the test effectively do nothing!

The intent was that with --check, the real pathnames should appear
to help debugging, e.g.
    /var/lib/pcp/pmns/root ... BAD
but without --check, the pathnames should include the PCP_* variables
so that the output is deterministic
    $PCP_VAR_DIR/pmns/root ... BAD

There was a related issue with directories that are checked ...
without --check we should emit only a line for the directory name if
the contents are all OK, else emit a line for each failing file within
the directory.  Again, this makes the output deterministic across
different platforms where directory contents may not be the same,
and certainly may not be in the same order as reported by find(1).