qa/common.check: make default PMDA install optional in _restore_pmda_install()
The default PMDA install recipe:
    $ ./Install </dev/null
may not work if the installation is designed to be interactive, or
requires explicit input values.

To accommodate this, _restore_pmda_install() supports an optional 2nd
parameter and if this is not empty, the default PMDA install will not
be attempted.  It is assumed that the caller will be taking care of
PMDA's final well-being if this is the case.