pmlogger: change to type=notify, update service units and logutil scripts
Change pmlogger service over to type=notify. Update rc script to
support "start-systemd" and use that in the service unit. The
primary logger is the main pid and we invoke it from pmlogger_check
with the new -N flag to notify the service manager (if any) when
the service has started and is ready.

Also subtly change how pmlogger(s) are launched from pmlogger_check
so they are an immediate child of init (aka systemd), so we get
rid of the dreaded but benign journal messages about not being a
child of systemd., etc.

Boot time testing shows the pmlogger primary pid starts in ~ 1 to 2s.
The remaining 8 to 10 seconds in boot time is spent in pmlogconf, for
which we have plans to improve. Until those improvements are available,
users can set PMLOGGER_CHECK_SKIP_LOGCONF=yes in /etc/sysconfig/pmlogger
once their generated pmlogger configs are stable.

Almost all existing QA in groups logutil, pmlogger and pmlc are still
passing (a few tweaks coming in later commits, including an SELinux
policy update).