pmlogconf: rewrite the shell scripts as a single C program for speed
The venerable pmlogconf utility has been rewritten as a program now,
to resolve performance issues around its use in pmlogger setup.  Its
now able to complete a typical localhost configuration in 20-30 msec
where the previous version took 6-8 sec to run.  This should help us
with both local service startup times, and also scaling up pmlogger
farms to greater numbers of hosts.

The functionality is equivalent, with the exception of retiring the
IRIX v1 format; for the past 10+ years this format was automatically
transitioned to the v2 format used in all open source releases, and
it just doesn't seem worthwhile carrying that code forward anymore
(it would need to be re-implemented).

One small change is that the groups are now kept sorted, in-memory
and on-disk - this sometimes results in a one-time reconfiguration,
and several of the QA tests have slightly different output.

The (undocumented, internal) pmlogconf-setup script has been retired
completely.  Previously this functionality was implemented by these
two scripts, now its an all-on-one command.  I've added some notes
to the pcpcompat(1) man page describing this, and a -s option to the
new pmlogconf in case anyone needs that old output format (I don't
expect this to be the case, however it is handy for use by QA).

A new -g/--group option is added, allowing an existing setup to be
queried in terms of states of individual groups - this is expected
to be used in configuration automation systems (like Ansible).

Resolves https://github.com/performancecopilot/pcp/issues/895