pmlogconf, pmdapmcd: mechanism for detecting PMDAs without static metrics
Some agents (e.g. netcheck) do not provide any static metrics - everything
exported is based on agents configuration.  For such cases, a pmcd metric
is added to allow metric-value-based detection of installed agents.  The
pmlogconf configuration file for netcheck is added to make use of it, and
regression test qa/1692 exercises it.