pmie,pmlogger_check,daily: set systemd working directory
We've had a report of failure in pmie_daily related to attempting
to set cwd to an unreadable directory (/root).  Although not able
to reproduce it locally, I believe the problem can be resolved by
forcing systemd to set the working directory to pcp user accounts
homedir (/var/lib/pcp) which is always readable by pcp.

Resolves Red Hat BZ #1832262.