pmlogrewrite: warn and do nothing for empty archives
Note this is for empty archives (not empty files), where the
archive contains a label record and no data records.

In some cases, the previous behaviour (to treat this as an error
when the pmNewContext() failed) would cause pmlogger_daily to fail
in the log rewriting step, which stopped the merging.

Included in the commit is a small man page change and a new QA
test (1195).