pcp-atop: ensure the dereferenced task pointers are validated
Corrections to the NULL pointer checks in task comparison routines,
ensuring we always verify task pointers, not pointers to pointers.

Still unable to reproduce this reliably, but local testing of the
atop group suggests all is well and poring over customer valgrind
output suggests this might be the final solution.

Related to Red Hat BZ #1818710