qa: refactor test 510 for speed and correctness
Recently qa/510 started failing with systemd path changes.
It was taking 160+ seconds to fail - this was largely due to
the copious amounts of lock-related libpcp diagnostics we're
now generating under -Dall.  Found the reason for -Dall to be
triggering a __pmDumpResult which can also be triggerd using
-Dlog.  Switched that over, and also switched over to common
QA systemd / cron job control functions.

Test passing again, controls cron and systemd more correctly,
and runs in 12sec.