libpcp: fix hostspec unparsing to deal with IPv6 bracketing
There is a convention in IPv6-aware tooling to square-bracket
enclose host addresses, as they contain colons which can be a
source of confustion when followed by a port name.  This is a
common practice in PCP IPv6 hostnames.

When we 'unparse' a hostname in libpcp (after splitting it on
proxy boundaries, etc) we were not checking whether previous
hostnames required braces (contain ':' characters) which was
resulting in parsing errors down the track.

Fix this, to detect when unparsing must put square brackets
back - and make it so.  Test 719 is extended to exercise this
case.