pmrep: fix label handling for writing archives in pmrep and pcp2xxx
Following on from Marko's work, fix the arguments to pmiPutLabel(3)
and a few other bugs. pmrep and associated pcp2xxx tools now correctly
read/write labels to/from archives as well as for live reporting,
when used with the new -m flag.

A new QA test is still coming, and python2 support needs checking,
but this feature is now looking pretty healthy.