qa/1793: new test exercising labels support for pmrep and pcp2xxx
Runs pmrep with the new -m flag to record an archive with labels
(this uses pmiPutLabel from the python binding for LOGIMPORT(3)
under the hood). Then examines the archive with pmdumplog -e and
then replays it with pmrep, pcp2csv and pcp2json.

Related https://github.com/performancecopilot/pcp/issues/806