pmie service: fix systemd race where rc script exits before daemon writes pidfile
If the service daemon hasn't written it's pidfile by the time the rc script exits,
some versions of systemd consider the service to have failed. Delay the rc script
a bit to allow the service daemon to get started before it is re-parented to
systemd.

Tested qa/115 and qa/1297 (with one and then two CPUs) -
  f32, systemd-245.4-1.fc32.x86_64
  rhel78 systemd-219-73.el7_8.6.x86_64
  centos77 systemd-219-73.el7_8.6.x86_64

Needs more testing on other platforms, especially with only one CPU and older
versions of systemd.