configure.ac: Force man pages for Debian-style builds
The dpkg packaging rules expect the man pages to be present, but our
"smart" configure logic for figuring out what style of man pages
are required was failing on container-based builds in CI (presumably
because there are no man pages installed there).

Force the inclusion of the man pages by setting the relevant variables
if the "smart" logic fails to set them.

Also picked up that configure was not remade after last configure.ac
change, so there are some changes in configure that are not caused
by the configure.ac changes in this commit.