src/pmlogger/pmlogger_check.sh: fix locking snarfoo
In the (unlikely) case that pmcd is not running and we're trying
to start the primary pmlogger, there was a serious bug where the
script exited and left the primary pmlogger lock in place.

The fix is to not exit, report an error, skip the primary pmlogger
and keep going.