pmsearch: full text search for metrics, instances and help
Uses the existing archive discovery in pmproxy for metadata
(incl help text) and the RediSearch module for Redis to get
a responsive index over PCP text metadata.

A command line tool - pmsearch - is added, as is a REST API
for web applications and grafana-pcp.

Further docs, QA and additional features to come but this is
enough to start boot-strapping the user interface work.

Relates to https://github.com/performancecopilot/pcp/issues/950