src/pmlogger/utilproc.sh: make _save_prev_file() less chatty
This helper function is intended to save <somefile> to <somefile>.prev.

It is used in the pm*_check and pm*_daily scripts for their diagnostic
log files.

But there is no locking in place, so we're exposed to race conditions.
It does not really matter (locking would be overkill) because if the
same script is being run often enough to hit the race window then the
contents of <somefile> and <somefile>.prev are like buying a ticket
in the lottery (you're clueless and have no chance of winning).

This commit silences _save_prev_file() when rm or cp fail, which is
a possible result of the race.

It has become more visible because pmlogctl is able to launch
pmlogger_check much faster than it was launced in the past, at least
in the qa tests.