libpcp, pmlogpaste, libpcp_import: fixes to labels setting functionality
Add guards to libpcp and libpcp_import functions that deal with context
level labels, such that the ident passed in is irrelevent (as it should
be) - previously we assumed this would be passed as PM_ID_NULL, but any
number (eg zero) could seem reasonable to a developer using the APIs.

Update pmlogpaste to make this a bit more clear, and qa/1821 now works.