src/pmie/pmie.service.in: pmie needs to be Before not After
If pmie_check.timer (and to a lesser extent pmie_daily.timer) run
before pmie then the we have a race to start pmie processes multiple
times, and this may lead to a protocol failure for the pmie service
which is left in a bad state.

Same problem already fixed for pmlogger some time ago.