Assorted systemd oneshot services: use KillMode=none
For all our *_check and *_daily scripts, when they are finished we
do NOT want systemd sending SIGTERM to any continuing processes that
systemd thinks belong to the associated control groups.

This change stops the infanticide via SIGTERM of pmie and pmlogger
processes started from the pmie, pmlogger and pmfind services (and/or
their associated .timer services).