src/pmie/pmie_check.*: make pmie_check PartOf pmie
When pmie is stopped by systemctl we need to also stop both the
pmie_check.service and pmie_check.timer.  If this does not happen
the next start pmie may report errors due to a race between
pmie.service and pmie_check.timer to get the primary pmie running
(pmie.service needs to win this race).