jumbo man page changes
First off the man/Check script has been fed steriods and now detects
a range of badness in the man pages mostly around consistency,
correctness and interaction with man(1), especially man -k ...
and man <section#> tool-or-function-name.

Also the debian packaging is checked for the section 3 entries that
are NOT in the pcp package.

And there are a bunch of man page changes to ensure man/Check passes
without errors or warnings.