src/include/builddefs.in: re-order -I options in CFLAGS
This was causing a build breakage on *BSD where libuv exists and
the headers for libuv are in /usr/pkg/include which is where the PCP
headers end up being installed also.

This change moves the build relative -I options to be before LCFLAGS
in CFLAGS, so our headers are inspected before /usr/pkg/include
(which is specified via LCFLAGS if needed for libuv, for example).

Otherwise the build uses the installed PCP headers, and when the
headers change (like adding a debug flag), the build will fail.