qa/admin/check-vm: fix Python version heuristic
Without -f, check-vm uses a "guess which version of XYZ you should
be using" heuristic, for XYZ being Python and Qt.  The Python bit
was wrong and was excluding python3-all-dev from the manifest when
version 3 was guessed for Python.

Was causing a CI build breakage for the ubuntu2004-dierct platform.