src/include/builddefs.in: tweak INSTALL_MAN
We really don't need the sort here, and it leads to some slightly
odd relationships between the installed files and the symlinks, e.g.
for pmdacisco with the sort
	parse.1 is a file
	pmdacisco.1 -> parse.1
	probe.1 -> parse.1
without the sort
	pmdacisco.1 is a file
	parse.1 -> pmdacisco.1
	probe.1 -> pmdacisco.1

There is of course another whole can of worms about these specific
man pages, but that's another topic.