pmview: fix GNUmakefile for bison 3.7
gram.c includes the following comment on bison < 3.7:

    In a future release of Bison, this section will be replaced by
    #include "gram.tab.h"

In bison 3.7 this #include is here, however the GNUmakefile renames
gram.tab.h to gram.h, so the required gram.tab.h cannot be found, which
breaks the build. This change instructs bison to output gram.c and
gram.h directly (instead of outputting gram.tab.h and renaming this file
in the GNUmakefile), so it includes the correct location to the header file.