libpcp: fortify pmns code for gcc10 warnings
gcc10 was warning about a potential call to malloc(0) and then an
invalid consequential assignment. Fortify the argument checking
in backname() in pmns.c to avoid malloc(0) and instead return
PM_ERR_PMNS. In practice there doesn't seem to be anywhere calling
this internal libpcp function with such invalid arguments.

QA for group 'pmns' was checked and still passes.