src/libpcp: derived metric changes for register after open
Previously, we had (incorrectly) assumed that the set of global derived
metrics would no change for a context after the pmNewContext() ... but
if a subsequent pmRegisterDerived*() is performed, we need those newly
defined derived metrics to be visible, not just in future contexts,
but also in all currently open contexts.

This is precursor work to the refactoring needed to support per-context
derived metrics alongside the (existing) global derived metrics.