qa/1886: update and extend pmseries query tests
Added additional query tests for arithmetic expressions
(for +, -, * and /), round() and a few others.

Also changed the arguments to pmseries to specify -Z UTC
so the dates reported in the pmseries output for each test
are the same no matter what the local timezone is.

Also remade 1886.out with the above changes.