pmdadm: remove forced_restart to avoid pmcd rc script recursion
The DM pmda needs to run as root, which is why it had the
forced_restart=true variable setting in it's Install script.
When the Install script was launched by the pmcd rc script
(due to .NeedInstall from a virgin install or upgrade involving
pcp-zeroconf), the pmcd service would be restarted (recursively)
so it could launch pmdadm as root before dropping privileges.

Since we now have pmdaroot as the parent of the DM PMDA, it
is no longer necessary to restart pmcd and hence we can avoid
the rc script recursion that has been causing issues with
systemd, CI and other container installations.

THis also applies to the postfix PMDA.

Have tested and verified pmda.dm and pmda.postfix QA groups.
Have tested upgrades and virgin installs (on f32). Let's see how
it goes with CI where systemd in the containerized environment
seems to be more sensitive to the rc script recursion.

There may be some QA fallout with tests involving error injection.

Resolves https://github.com/performancecopilot/pcp/issues/1001